Removed samaritan folder. Imported navbar to all pages
